added views

+++ InvoiceingProduct/InvoiceingProduct/Models/DBObjects/Invoice.cs
@@ -13,8 +13,8 @@ namespace InvoiceingProduct.Models.DBObjects
         public Guid IdInvoice { get; set; }
         public Guid IdPurchase { get; set; }
         public DateTime InvoiceDate { get; set; }
-        public int InvoiceAmount { get; set; }
-        public int TaxAmount { get; set; }
+        public decimal InvoiceAmount { get; set; }
+        public decimal TaxAmount { get; set; }
         public DateTime DueDate { get; set; }
         public string? Comments { get; set; }
